+++ Tests/Functional/Controller/CrawlerControllerTest.php
@@ -26,7 +26,6 @@ use AOE\Crawler\Tests\Functional\BackendRequestTestTrait;
 use AOE\Crawler\Value\QueueFilter;
 use PHPUnit\Framework\MockObject\MockObject;
 use TYPO3\CMS\Core\Authentication\BackendUserAuthentication;
-use TYPO3\CMS\Core\Information\Typo3Version;
 use TYPO3\TestingFramework\Core\AccessibleObjectInterface;
 use TYPO3\TestingFramework\Core\Functional\FunctionalTestCase;
 
@@ -102,7 +101,6 @@ class CrawlerControllerTest extends FunctionalTestCase
         bool $registerQueueEntriesInternallyOnly,
         bool $expected
     ): void {
-        (new Typo3Version())->getMajorVersion();
         $mockedQueueRepository = $this->getAccessibleMock(
             QueueRepository::class,
             ['getDuplicateQueueItemsIfExists']